improved the efficiency to mlogn and rectified code
